chore(Plugins): Fixing raknet builds

+++ Source/Utils/_FindFirst.h
@@ -47,7 +47,7 @@ typedef struct _findinfo_t
 } _findinfo;
 
 long _findfirst(const char *name, _finddata_t *f);
-int _findnext(intptr_t h, _finddata_t *f);
+int _findnext(long h, _finddata_t *f);
 int _findclose(intptr_t h);
 
 #endif
